feat: Implement carbon footprint calculation, scheduling, and management with a dedicated admin UI.

+++ backend/src/waste/entities/waste-material.entity.ts
@@ -20,7 +20,7 @@ export class WasteMaterial {
   @Column({ type: 'varchar', length: 255 })
   name: string;
 
-  @Column({ type: 'int', nullable: true })
+  @Column({ type: 'float', nullable: true })
   emission_factor: number;
 
   @Column({ type: 'varchar', length: 255, nullable: true })
